backend/python/mytts/requirements.txt
grpcio
grpcio-tools
protobuf>=3.19.0

# Bibliothèques TTS (choisir selon vos besoins)
pyttsx3  # TTS hors ligne
gtts     # Google Text-to-Speech (nécessite Internet)

# Optionnel: autres dépendances selon le moteur TTS utilisé
pygame   # Pour la lecture audio (optionnel)
